Impute missing values in a LipidomicsExperiment

Impute missing values in a LipidomicsExperiment

impute_na(
  data,
  measure = "Area",
  method = c("knn", "svd", "mle", "QRILC", "minDet", "minProb", "zero"),
  ...
)

Arguments

data

LipidomicsExperiment object.

measure

Which measure to use as intensity, usually Area, Area Normalized or Height. Default is Area.

method

The imputation method to use. All methods are wrappers for imputeLCMD package. These include

knn Wraps imputeLCMD::impute.wrapper.KNN(). Default. This requires an additional argument K (Number of neighbors used to infer the missing data).

svd Wraps imputeLCMD::impute.wrapper.SVD(). This requires an additional argument K (Number of principal components to use).

mle Wraps imputeLCMD::impute.wrapper.MLE(),

minDet Wraps imputeLCMD::impute.MinDet(),

minProb Wraps imputeLCMD::impute.MinProb(),

zero Wraps imputeLCMD::impute.ZERO(),

...

Other arguments passed to the imputation method.

Value

LipidomicsExperiment object with missing values imputed.

Examples

data(data_normalized)

# Replace with values calculated using K-nearest neighbors
impute_na(data_normalized, "Area", "knn", 10)
#> class: LipidomicsExperiment 
#> dim: 278 56 
#> metadata(2): summarized dimnames
#> assays(3): Retention Time Area Background
#> rownames(278): 1 2 ... 277 278
#> rowData names(24): filename Molecule ... total_cs Class
#> colnames(56): S1A S2A ... TQC_11 TQC_12
#> colData names(3): group Diet BileAcid

# Replace with values calculated from the first K principal components
impute_na(data_normalized, "Area", "svd", 3)
#> class: LipidomicsExperiment 
#> dim: 278 56 
#> metadata(2): summarized dimnames
#> assays(3): Retention Time Area Background
#> rownames(278): 1 2 ... 277 278
#> rowData names(24): filename Molecule ... total_cs Class
#> colnames(56): S1A S2A ... TQC_11 TQC_12
#> colData names(3): group Diet BileAcid

# Replace with Maximum likelihood estimates
impute_na(data_normalized, "Area", "mle")
#> class: LipidomicsExperiment 
#> dim: 278 56 
#> metadata(2): summarized dimnames
#> assays(3): Retention Time Area Background
#> rownames(278): 1 2 ... 277 278
#> rowData names(24): filename Molecule ... total_cs Class
#> colnames(56): S1A S2A ... TQC_11 TQC_12
#> colData names(3): group Diet BileAcid

# Replace with randomly drawn values from a truncated distribution
impute_na(data_normalized, "Area", "QRILC")
#> class: LipidomicsExperiment 
#> dim: 278 56 
#> metadata(2): summarized dimnames
#> assays(3): Retention Time Area Background
#> rownames(278): 1 2 ... 277 278
#> rowData names(24): filename Molecule ... total_cs Class
#> colnames(56): S1A S2A ... TQC_11 TQC_12
#> colData names(3): group Diet BileAcid

# Replace with a minimal value
impute_na(data_normalized, "Area", "minDet")
#> class: LipidomicsExperiment 
#> dim: 278 56 
#> metadata(2): summarized dimnames
#> assays(3): Retention Time Area Background
#> rownames(278): 1 2 ... 277 278
#> rowData names(24): filename Molecule ... total_cs Class
#> colnames(56): S1A S2A ... TQC_11 TQC_12
#> colData names(3): group Diet BileAcid

# Replace with randomly drawn values from a Gaussian distribution
# cerntered around a minimal value
impute_na(data_normalized, "Area", "minProb")
#> [1] 0.4174013
#> class: LipidomicsExperiment 
#> dim: 278 56 
#> metadata(2): summarized dimnames
#> assays(3): Retention Time Area Background
#> rownames(278): 1 2 ... 277 278
#> rowData names(24): filename Molecule ... total_cs Class
#> colnames(56): S1A S2A ... TQC_11 TQC_12
#> colData names(3): group Diet BileAcid

# Replace with zero (not recommended)
impute_na(data_normalized, "Area", "zero")
#> class: LipidomicsExperiment 
#> dim: 278 56 
#> metadata(2): summarized dimnames
#> assays(3): Retention Time Area Background
#> rownames(278): 1 2 ... 277 278
#> rowData names(24): filename Molecule ... total_cs Class
#> colnames(56): S1A S2A ... TQC_11 TQC_12
#> colData names(3): group Diet BileAcid