Description of lipidr datasets

lipidr-package has 3 datasets:

data_normalized Example lipidomics dataset, normalized & log2-transformed.

lipidDefaults A list of default mappings and annotations for lipids.

lipidnames_pattern A list of patterns used in parsing lipid names.

See below for detailed description of each dataset.

See also

Examples

data(data_normalized)