Metabolomics Workbench integration

These functions use Metabolomics Workbench REST API to support data mining of publicly available lipidomics datasets.

list_mw_studies(keyword = "lipid")

fetch_mw_study(study_id)

read_mwTab(mwTab)

read_mw_datamatrix(file)

Arguments

keyword
A keyword to search for in Metabolomics Workbench studies.

study_id
The Metabolomics Workbench study ID.

mwTab
File path or url for a mwTab file.

file
File path or url for the file containing the data matrix.

Value

list_mw_studies returns a data frame with studies matching the keyword. Study ID, title, author and details are retrieved.

All other functions return a LipidomicsExperiment object containing clinical and lipid intensity data.

Functions

list_mw_studies: retrieves a list of lipidomics studies from Metabolomics Workbench.

fetch_mw_study: downloads and parse full data for a study from Metabolomics Workbench using a study ID. The function returns a LipidomicsExperiment where users can directly apply lipidr analysis workflow.

read_mwTab: parses mwTab file into a LipidomicsExperiment.

read_mw_datamatrix: parses a Metabolomics Workbench data matrix into a LipidomicsExperiment. Data matrix downloaded from Metabolomics Workbench are parsed into a LipidomicsExperiment object to enable lipidr workflow analysis.

Examples

if (FALSE) {
list_mw_studies()
}
if (FALSE) {
fetch_mw_study("ST001111")
}